20260622dy1

+++ project/board/view.php
@@ -1,5 +1,17 @@
 <?php
+
+    session_start();
     header('Content-Type:text/html; charset=utf-8');
+
+    if(!isset($_SESSION['user_no'])){
+        echo "
+            <script>
+                alert('로그인이 필요합니다.');
+                location.href='../index.html';
+            </script>
+        ";
+        exit;
+        }
 
     $db = mysqli_connect('localhost','monster2026aix','a1s2d3f4!','monster2026aix');
 
@@ -37,7 +49,7 @@
 
             <nav class="main_nav">
                 <a href="./mypage.html">MY</a>
-                <a href="../index.html">로그아웃</a>
+                <a href="../backend/logout.php">로그아웃</a>
             </nav>
         </header>
 
@@ -79,8 +91,12 @@
             <p class="noact">등록된 댓글이 없으며 등록되지 않습니다 미구현</p>
 
             <div class="view_btn_box">
-                <a href="./edit.php?no=<?php echo $row['no']; ?>">수정</a>
-                <a href="../backend/delete.php?no=<?php echo $row['no']; ?>" onclick="return confirm('정말 삭제하시겠습니까?');"> 삭제</a>
+
+                <?php if($_SESSION['user_no'] == $row['user_no']){ ?>
+                    <a href="./edit.php?no=<?php echo $row['no']; ?>">수정</a>
+                    <a href="../backend/delete.php?no=<?php echo $row['no']; ?>" onclick="return confirm('정말 삭제하시겠습니까?');">삭제</a>
+                <?php } ?>
+
                 <a href="./board.php">목록으로</a>
             </div>
